requirements.txt
pandas==2.2.1
numpy==1.26.4
torch==2.2.1
transformers==4.38.2
nrclex==4.0
vaderSentiment==3.3.2
ruptures==1.1.9
scikit-learn==1.4.1.post1
matplotlib==3.8.3
seaborn==0.13.2
tqdm==4.66.2
python-dotenv==1.0.1
